Cleanup some code; Add some skeleton code
Removed all of the background and nametable init code.  Instead, just
turn off the background and fill the background palettes with $0F.

Added routines to flip sprite directions both vertically and
horizontally.

Added empty routines to check for collision with the four edges of the
screen.

Added a routine to cycle the first sprite palette using palettes stored
in the ROM.

Added some misc comments noting the function of some blocks of code.